Create a base Component class and restructure files
- Create a `Component` class to be extended by all components.
  - Attach shadow DOM and fill it with the template that was received.
  - Reset component styles without having to import to each component's individual stylesheet by applying it in the `Component` class.
