+++ src/components/MainFooter.js
@@ -1,3 +1,5 @@
+import Component from "./common/Component.js";
+
 const template = document.createElement("template");
 template.innerHTML = `
   <footer>
@@ -69,11 +71,9 @@ template.innerHTML = `
   <link rel="stylesheet" href="src/styles/components/MainFooter.css">
 `;
 
-class MainFooter extends HTMLElement {
+class MainFooter extends Component {
   constructor() {
-    super();
-    const shadowRoot = this.attachShadow({ mode: "open" });
-    shadowRoot.append(template.content.cloneNode(true));
+    super(template);
   }
 }
