+++ src/styles/components/AutocompletePanel.scss
@@ -1,4 +1,3 @@
-@use "../config/component-reset.scss" as *;
 @use "../config/design-system.scss" as *;
 
 :host {
