+++ src/styles/components/InfiniteCarousel.scss
@@ -1,4 +1,3 @@
-@use "../config/component-reset.scss";
 @use "../config/design-system.scss" as *;
 
 :host {
