+++ src/styles/components/shared/ToolTip.scss
@@ -1,7 +1,7 @@
-@use "../../config/component-reset.scss";
 @use "../../config/design-system.scss" as *;
 
 :host {
+  padding: 16px;
   display: block;
   position: absolute;
   top: 45px;
