+++ src/styles/config/reset.scss
@@ -1,5 +1,3 @@
-@use "./design-system.scss" as *;
-
 *,
 *::before,
 *::after {
